refactor(shorebird_cli): move apple classes into separate files
