+++ packages/shorebird_cli/lib/src/platform/apple/apple_platform.dart
@@ -1,0 +1,9 @@
+/// Apple-specific platform options, corresponding to different Flutter target
+/// platforms.
+enum ApplePlatform {
+  /// iOS
+  ios,
+
+  /// macOS
+  macos,
+}
