+++ packages/shorebird_cli/lib/src/platform/apple/invalid_export_options_plist_exception.dart
@@ -1,0 +1,13 @@
+/// {@template invalid_export_options_plist_exception}
+/// Thrown when an invalid export options plist is provided.
+/// {@endtemplate}
+class InvalidExportOptionsPlistException implements Exception {
+  /// {@macro invalid_export_options_plist_exception}
+  InvalidExportOptionsPlistException(this.message);
+
+  /// An explanation of this exception.
+  final String message;
+
+  @override
+  String toString() => message;
+}
